# swagger/swagger.yml
openapi: 3.0.0
info:
  description: This API provides a number of methods to retrieve historical data across
    multiple dimensions of the TON blockchain.
  title: TON Blockchain History APIs
  version: 0.0.1
  x-logo:
    url: https://tengram.com/TENGRAM_logo.svg
    altText: Tengram
...
servers:
  - url: https://api.tengram.com/
components:
  schemas:
    uint64:
      type: integer
      format: int64
      minimum: 0
      maximum: 18446744073709551615
...
    BlocksFeedResponse:
      type: object
      properties:
        scroll_id:
          type: string
        blocks:
          type: array
          items:
            $ref: "#/components/schemas/BlockInFeed"
...
    BlockInFeed:
      description: Block in feed
      type: object
      required:
        - workchain_id
        - shard
        - seq_no
        - time
        - start_lt
        - total_fees_nanograms
        - root_hash
        - file_hash
        - trx_count
        - value_nanograms
        - ihr_fee_nanograms
        - import_fee_nanograms
        - fwd_fee_nanograms
      properties:
        workchain_id:
          type: integer
          format: int32
        shard:
          $ref: "#/components/schemas/uint64"
        seq_no:
          $ref: "#/components/schemas/uint64"
        time:
          $ref: "#/components/schemas/uint64"
        start_lt:
          $ref: "#/components/schemas/uint64"
        total_fees_nanograms:
          $ref: "#/components/schemas/uint64"
        root_hash:
          type: string
        file_hash:
          type: string
        trx_count:
          type: integer
          format: int64
        value_nanograms:
          $ref: "#/components/schemas/uint64"
...
    AccountWhale:
      description: ""
      type: object
      required:
        - addr_raw
        - addr_uf
        - balance_nanogram
        - balance_percentage_of_total
      properties:
        addr_raw:
          type: string
          description: Raw whale account address.
        addr_uf:
          type: string
          description: User-friendly whale account address.
        balance_nanogram:
          description: Account balance in nanograms.
          $ref: "#/components/schemas/uint64"
        balance_percentage_of_total:
          type: number
          format: double
          description: Ratio of this accounts balance to total nanograms.
    AddrTopByMessageCountResponse:
      description: ""
      type: object
      properties:
        top_in:
          type: array
          items:
            $ref: "#/components/schemas/AddrCount"
        top_out:
          type: array
          items:
            $ref: "#/components/schemas/AddrCount"
    AddrCount:
      description: ""
      type: object
      required:
        - addr
        - addr_uf
      properties:
        workchain_id:
          type: integer
          format: int32
        addr:
          type: string
        addr_uf:
          type: string
          description: User-friendly account address.
        count:
          type: integer
          format: int64
    MessageFeedResponse:
      type: object
      required:
        - scroll_id
        - messages
      properties:
        scroll_id:
          type: string
        messages:
          type: array
          items:
            $ref: "#/components/schemas/MessageFeed"
...
    MessageFeed:
      description: ""
      type: object
      required:
        - workchain_id
        - shard
        - seq_no
        - lt
        - time
        - message_lt
        - trx_hash
        - direction
        - src_workchain_id
        - src
        - src_uf
        - dest_workchain_id
        - dest
        - dest_uf
        - value_nanogram
        - total_fee_nanogram
        - bounce
        - body
      properties:
        workchain_id:
          type: integer
          format: int32
        shard:
          $ref: "#/components/schemas/uint64"
        seq_no:
          $ref: "#/components/schemas/uint64"
        lt:
          $ref: "#/components/schemas/uint64"
        time:
          $ref: "#/components/schemas/uint64"
        message_lt:
          $ref: "#/components/schemas/uint64"
        trx_hash:
          type: string
        direction:
          type: string
        src_workchain_id:
          type: integer
          format: int32
        src:
          type: string
        src_uf:
          type: string
          description: User-friendly source account address.
        dest_workchain_id:
          type: integer
          format: int32
        dest:
          type: string
        dest_uf:
          type: string
          description: User-friendly destination account address.
        value_nanogram:
          $ref: "#/components/schemas/uint64"
        total_fee_nanogram:
          $ref: "#/components/schemas/uint64"
        bounce:
          type: boolean
        body:
          type: string
...
    MessagesByTypeResult:
      description: ""
      type: object
      required:
        - rows
      properties:
        rows:
          type: array
          items:
            $ref: "#/components/schemas/MessagesByTypeTimeseries"
    MessagesByTypeTimeseries:
      type: object
      required:
        - workchain_id
        - type
        - time
        - messages_count
      properties:
        workchain_id:
          description: workchain_id
          type: integer
          format: int32
        type:
          description: type
          type: string
        msg_type:
          description: Msg type
          type: string
        time:
          $ref: "#/components/schemas/uint64Array"
        messages_count:
          $ref: "#/components/schemas/uint64Array"
    VolumeByGramsResult:
      description: ""
      type: object
      required:
        - rows
      properties:
        rows:
          type: array
          items:
            $ref: "#/components/schemas/VolumeByGramsTimeseries"
    VolumeByGramsTimeseries:
      type: object
      required:
        - workchain_id
        - time
        - volume_grams
      properties:
        workchain_id:
          description: workchain_id
          type: integer
          format: int32
        time:
          $ref: "#/components/schemas/uint64Array"
        volume_grams:
          $ref: "#/components/schemas/stringArray"
    MessagesOrdCountResult:
      description: ""
      type: object
      required:
        - rows
      properties:
        rows:
          type: array
          items:
            $ref: "#/components/schemas/MessagesOrdCountTimeseries"
    MessagesOrdCountTimeseries:
      type: object
      required:
        - workchain_id
        - time
        - count
      properties:
        workchain_id:
          description: workchain_id
          type: integer
          format: int32
        time:
          $ref: "#/components/schemas/uint64Array"
        count:
          $ref: "#/components/schemas/uint64Array"
    TimeseriesBlocksByWorkchain:
      description: ""
      type: object
      required:
        - time
        - masterchain
        - workchain0
      properties:
        time:
          $ref: "#/components/schemas/uint64Array"
        masterchain:
          $ref: "#/components/schemas/uint64Array"
        workchain0:
          $ref: "#/components/schemas/uint64Array"
...
    stringArray:
      description: String array type
      type: array
      items:
        type: string
...
    uint64Array:
      description: Uint64 array type
      type: array
      items:
        $ref: "#/components/schemas/uint64"
...
    error:
      type: object
      required:
        - message
      properties:
        error:
          type: boolean
          description: Error
        message:
          type: string
          description: Error message
paths:
...
  /v1/blocks/feed:
    get:
      summary: Get blocks feed
      tags:
        - core
      description: >
        Get blocks feed <br>
        This method is available by following paths: /v1/block/feed, /v1/blocks/feed, /v1/b/feed
      parameters:
        - name: scroll_id
          in: query
          description: Scroll ID
          schema:
            type: string
        - name: limit
          in: query
          description: Limit blocks
          schema:
            type: integer
            default: 30
        - name: workchain_id
          in: query
          description: Workchain Id filter
          schema:
            type: integer
            format: int32
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlocksFeedResponse"
        default:
          description: generic error response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/error"
  /v1/timeseries/blocks-by-workchain:
    get:
      summary: Blocks by workchain
      tags:
        - main
      description: ""
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TimeseriesBlocksByWorkchain"
        default:
          description: generic error response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/error"
  /v1/timeseries/messages-by-type:
    get:
      summary: Messages by type
      tags:
        - main
      description: ""
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessagesByTypeResult"
        default:
          description: generic error response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/error"
  /v1/timeseries/volume-by-grams:
    get:
      summary: Volume by grams
      tags:
        - main
      description: ""
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VolumeByGramsResult"
        default:
          description: generic error response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/error"
  /v1/timeseries/messages-ord-count:
    get:
      summary: Messages ord count
      tags:
        - main
      description: ""
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessagesOrdCountResult"
        default:
          description: generic error response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/error"
  /v1/messages/feed:
    get:
      summary: Messages feed
      tags:
        - core
      description: ""
      parameters:
        - name: scroll_id
          in: query
          description: Scroll ID
          schema:
            type: string
        - name: limit
          in: query
          description: Limit messages
          schema:
            type: integer
            default: 50
        - name: workchain_id
          in: query
          description: Workchain Id filter
          schema:
            type: integer
            format: int32
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageFeedResponse"
        default:
          description: generic error response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/error"
...
  /v1/addr/top-by-message-count:
    get:
      summary: Top account by message in/out
      tags:
        - main
      description: ""
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AddrTopByMessageCountResponse"
        default:
          description: generic error response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/error"
  /v1/top/whales:
    get:
      summary: Top whales accounts by gram amount held
      tags:
        - main
      description: ""
      parameters:
        - name: workchain_id
          in: query
          description: Get top whales for specific WorckchainId. Do not specify to get
            workchain agnostic stats.
          schema:
            type: integer
            format: int32
        - name: limit
          in: query
          description: Get top whales with limit. Do not specify to get first 50 whales.
          schema:
            type: integer
            format: int64
            default: 50
        - name: offset
          in: query
          description: Get top whales with offset. Do not specify to get first whales (with no offset applied).
          schema:
            type: integer
            format: int64
            default: 0
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AccountWhale"
        default:
          description: generic error response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/error"
...